requirements.txt
django-cors-headers>=2.4.0,<2.5.0
djangorestframework-jwt>=1.11.0,<1.12.0
flake8>=3.6.0,<3.7.0
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
djangorestframework==3.11